Replaces the trasto gem with shioyama/mobility
Mobility has been getting a lot of attention recently and looks much
better architected than the moribund trasto. Interestingly, it allows
for queries based on the translated value which use the normal AR API.

With this commit we also switch to a jsonb column backend for translated
attributes, which allows us to remove the hstore extension and benefit
from structured data in translation much easier.