Fixes Question#requiring_response_from when no custom questions (#51)
The questions of a customized quiz which it inherits from the template
quiz still belong to the template quiz. As a result, looking up
answers_needed from the deployment with the question's quiz_id failed
for questions that the quiz inherited. There is no deployment of the
template quiz, only of its descendent. When there are any custom
questions at all, this falsey comparison with NULL is ignored, but
without, the Quiz#requires_response_from? would never return true.

To get around this, we're looking up the deployment by case and group.
Since it is an invariant (due to what we receive from LTI / the
ownership of quizzes by cases) that deployments.case_id is unique for
group_id, quiz_id pairs and that deployments.quiz_id is unique for
group_id, case_id pairs, we can be sure this will work.