Refactors notification to reply_notification
Notification had been designed as a catch-all pseudopolymorphic class,
switching on category and a data blob. In fact, all of that was betting
on the need for other kinds of notifications which for now we don’t
have. YAGNI. If we do, we can revisit what might be best as
polymorphism. But don’t forget that Rails polymorphic associations are
different 😞