Fixes to collapsing summary: editing and without metadata (#66)
The summary should never collapse if there is no other metadata to
display underneath it. In fact, the simplest solution is that the
summary never needs to be recollapsed after it is open.

Additionally, the summary can’t collapse while it is being edited, and
besides: something about the wrapper broke the styling of the
EditableText. Now we’ll just not use an InnerContainer when editing.