feat: rework “new Edgenote” workflow
We had been requiring editors to manually input Edgenote slugs when
attaching an Edgenote: a unique slug would create a new Edgenote, the
slug of another Edgenote on the case would reattach it, and a slug of a
different case’s Edgenote would... fail silently. This was always a
stop-gap.

This implements an interface for Edgenote creation and reattachment that
makes a lot more sense. Since we’re no longer using Edgenote slugs as
URL parameters, there’s no reason for editors to have to set them
manually: when there are no Edgenotes to reattach, the new Edgenote
button just automatically creates one. When there are Edgenotes that
have been deleted from one part of a case and could be reattached, we
stop and present a dialog offering the editor the choice to reattach or
create new.

To make this split workflow work, the toolbar button calls a function
that returns a promise of an Edgenote slug. When resolved, a
DraftEntity is created with that slug and added to the card. The
promise is resolved either by the POST request creating a new Edgenote
or by the user’s click on an Edgenote in the EdgenoteLibrary.