feat: implement method of calculating which instructors to prompt
After a professor has deployed a case in their classroom, we want to
invite them to write up a retrospective of their experience as a post in
the instructors-only community “CaseLog.”

We look for deployments whose members were much more active in the case
14 to 7 days ago than they were in the last 7 days. This is defined as
an 80% drop-off of use from wee to week, where the previous week
measured over 500 events. We exclude any deployments with fewer than
five readers as a sanity check.

These magic numbers were chosen empirically based on recorded data at an
early stage and should be subject to review.