fix: nullify limit in `part_of_catalog` scope
Because we can't select distinct unless we couple the select clause to 
the order set in `most_popular`, we are adding the two lists of ids and 
using a subquery. If we limit a relation that we later scope with this, 
the limits apply to the unordered, indistict query and limit the 
resulting tag list to very few. We can get around this by explicitly 
unsetting the limit in the subquery.