fix: fail safe by deleting attachments when validation fails
In Rails < 6, attachment changes are persisted as soon as they are
assigned, regardless of whether the record they’re attached to is
successfully saved. That’s why Rails < 6 doesn’t support validation of
attachments. We need to, but the model is left in an inconsistent state
if validation fails: the old blob is deleted as soon as the new
attachment is assigned, but if validation fails the old attachment is
preserved, and points to a dead blob.

Therefore, we’re monkey patching rails to proactively delete any
extant attachments that point to a blob before purging it. This will
only be necessary until Rails 6 comes out, because of
https://github.com/rails/rails/pull/33303