feat: allow cards to be moved to other pages or to new pages
Blocked on https://github.com/atlassian/react-beautiful-dnd/issues/930
for actual mouse-driven drag and drop (since the size of a card is too
large relative to the table of contents), but using the keyboard mode
this works. Drag a card onto a page in the TOC to move it to that page.
Or drag it between two entries in the TOC to create a new page at that
location.