feat: create SpotlightManager
We’ll have Spotlights all over the page, in and out of React apps,
identified by keys that the user’s onboarding script may include. We
only want one Spotlight visible at a time, which should be the one with
the highest priority that the user hasn’t acknowledged. The
SpotlightManager keeps track of all mounted Spotlights and controls
their visibility.

Co-Authored-By: Pearl Zeng <pearlzhuzeng@users.noreply.github.com>