test: make locking test more resiliant
For some reason relating to caching, in Chrome in development mode
there’s a flash of weird styles before the blueprint styles take effect.
If an editable text element measures itself before that flash is gone,
it can get stuck in too small a state for the whole title to be visible.
Then, we can’t find it to click on and the test fails.

The kicker is known to be shorter so it should not have this problem.