Focus should work as intended on all browsers, with invisible focus ring on NextLink
I added the styling to the application wide application.css which is for sure not the right place for it. It's working as intended right now but I'm not sure how the CSS is organized and where to put it so it doesn't affect the entire application. Other than that, focus works as intended in Chrome, Firefox, and Safari, although I needed to change settings in Firefox and Safari to get it to work.