Change Mapbox token in index.jsx
The token that was being used here had no URL restrictions and is exposed to the client, meaning it is possible to grab it from the Gala site and use it on another project. A sudden spike in usage this month led to my account having a non-zero bill (still only like $2) for the first time. I'm worried that an LLM may have grabbed the token and suggested it for reuse in someone else's project, so I'm swapping it here for a token with URL restrictions that limiting its use to `https://www.learngala.com`. Then I will rotate my public token